+++ History.md
@@ -1,4 +1,19 @@
 
+1.0.0 / 2017-07-14
+==================
+
+* compatible with socket.io 2.0.x
+* update engine.io-client
+* custom encoder/decoder support
+* fix socket id
+
+0.9.0 / 2017-07-11
+==================
+
+* compatible with socket.io 1.7.4
+* bump engine.io-client
+* send query on connect
+
 0.8.3 / 2016-12-12
 ==================